From-SVN: r245416

+++ gcc/testsuite/gcc.dg/graphite/scop-dsyr2k.c
@@ -1,8 +1,9 @@
+/* { dg-require-effective-target size32plus } */
 #define NMAX 3000
 
 static double a[NMAX][NMAX], b[NMAX][NMAX], c[NMAX][NMAX];
 
-void dsyr2k(long N) {
+void dsyr2k(int N) {
 	int i,j,k;
 	
 #pragma scop
@@ -16,5 +17,4 @@ void dsyr2k(long N) {
 #pragma endscop
 }
 
-/* { dg-final { scan-tree-dump-times "number of SCoPs: 1" 1 "graphite"} } */ 
-
+/* { dg-final { scan-tree-dump-times "number of SCoPs: 1" 1 "graphite" { xfail *-*-* } } } */
