+++ gcc/testsuite/gcc.dg/graphite/scop-dsyrk.c
@@ -1,9 +1,10 @@
+/* { dg-require-effective-target size32plus } */
 #define NMAX 3000
 #define MEASURE_TIME 1
 
 static double a[NMAX][NMAX], c[NMAX][NMAX];
 
-void dsyrk(long N) 
+void dsyrk(int N) 
 {
   int i,j,k;
 
@@ -18,4 +19,4 @@ void dsyrk(long N)
 #pragma endscop
 }
 
-/* { dg-final { scan-tree-dump-times "number of SCoPs: 1" 1 "graphite" } } */
+/* { dg-final { scan-tree-dump-times "number of SCoPs: 1" 1 "graphite" { xfail *-*-* } } } */
